Added 'manager/console/src/main/jnlp/template.vm_jnlp-signing' to make maven produce 'manager/console/target/APPLICATION_TEMPLATE.JNLP'. Using a second template.vm is only a workaround. The better solution would be to use only one jnlp-template-file that contains a property ${...} which gets replaced with "*" and "$$codebase" via Maven/'manager/console/pom.xml' in order to generate 'APPLICATION_TEMPLATE.JNLP' and 'launch.jnlp'.
These files are produced by the 'webstart-maven-plugin' which is configured in 'manager/console/pom.xml'.

The 'truezip-maven-plugin' (also configured in manager/application/pom.xml) is used to insert 'APPLICATION_TEMPLATE.JNLP'
into 'manager/console/target/console-${project.version}.jar/JNLP-INF/'.
That became necessary in order to conform to the oracle-security-policy/sercurity-baseline
introduced with java-7u45 which requires the jnlp-file to be signed.