Made some changes in the OpenVNCConnectionAction
Removed the NPE-catching and handled it with a if-statement.

Removed the variables case1 and case2.

Implemented a new way to get the client ip address from the client-log.
Now we get the address from the bottom to be sure that it is the newest
one. This way is now used for Linux and Windows.