Fix incompatibility with Linux NFS client
Since Kernel 5.7 sending a cookie=0 for the last file in a READDIR reply
causes listings over mounted directories to get stuck in an infinite
loop. Fixed by simply sending the same value as the Linux' NFS server.

(This is technically a bug in the Linux kernel as the cookie of the last
file should not have any effect on anything.)