Force higher version of sub-dependency
Some dependencies of glassfish jaxb-runtime are missing in some
repositories (for reasons?), resulting in failure on CI server
(i.e. com.sun.istack:istack-commons-runtime:jar:3.0.6 and
      com.sun.xml.fastinfoset:FastInfoset:jar:1.2.14)

Forcing a later build of the same version of this dependency of
hibernate-core, itself a dependency of spring-boot, seems to yield
more available sub-sub-sub-dependencies (hopefully).