Fix deadlock with package update event
After the package manager has un/installed package/s (in the same thread
as Vaadin) it publishes an `PackageEvent` which calls the event handler
in `ClientNotifier` (still in the same thread). The subsequent call to
`WebSocketSession.sendMessage` ends up deep inside undertow code where
it (sometimes) tries to acquire a resource still locked by the initial
HTTP request, resulting in a dead lock.

This fix simply starts the websocket logic in a new thread.