Fix "installation" of multiple versions of package
Make it impossible for the user to select more than one version of a
package for installation.

This prevents the package manager from fantasizing about installing
multiple versions of a package with identical file paths. Uninstalling
one of those versions removes the "shared" files, breaking all other
versions, that the package manager deems as installed.

To enable a partial multi-selection mode in a Vaadin grid, the Vaadin
control elements (namely the first column and checkboxes) are hidden
and the required functionality is (re-)implemented via click handler and
the `deduplicatePackageSelection` method.