Add workaround for buggy built-in LDAP server
The built-in LDAP server is very old (for reasons) and fails when
non-word characters are involved in searching. It also is unable to
understand escaped characters in filters. Since updating the LDAP server
(in the hope that a later version has fixed this) is too involved and
time-consuming for the moment (for reasons), this commit adds a clunky
workaround for this issue.

The workaround is used (only) with searches that involve names/DNs that
may contain "special" characters.

(Apologies for future readers of the code.)