Fix removal of applications from application group
If applications are removed and other applications are added to an
application group, the removal does not actually happen.

This fixes the removal by removing the application items from the
"abstract" application groups (instead of just removing the groups from
the applications), analogous to the addition of applications (which also
explicitly adds the applications to the groups).