Fix/extend workaround for LDAP's uniqueMemberMatch
Besides neither handling "special" character nor escaped characters, the
uniqueMemberMatch of the built-in LDAP server also wrongly compares DNs
case-sensitive (which is not helped by the woraround itself doing the
same).

This completly rewrites the workaround to be simpler (and even less
efficient) and more correct by doing the filtering in Java code.