Add support for usergroups in some REST endpoints
Get usergroups within REST request for all user-assignable items (i.e.
applications and printers).

If secondary LDAP is active, the client acquires the groups of the user
on login, instead of the server caching usergroups or acquiring them for
each REST request. This is done primarily to maintain REST performance
(and for a much easier implementation).